requirements.txt
# ─── Web Framework ────────────────────────────────────────────────────────────
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ─── Database ─────────────────────────────────────────────────────────────────
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
pgvector==0.2.5
alembic==1.13.1

# ─── ML / Vision ──────────────────────────────────────────────────────────────
insightface==0.7.3
onnxruntime==1.17.1          # CPU inference for insightface
torchreid==1.4.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python==4.9.0.80
deep-sort-realtime==1.3.2

# ─── Utilities ────────────────────────────────────────────────────────────────
numpy==1.26.4
pillow==10.3.0
pydantic==2.7.0
pydantic-settings==2.2.1
python-dotenv==1.0.1
requests==2.31.0
scipy==1.13.0

# ─── Dev / Testing ────────────────────────────────────────────────────────────
httpx==0.27.0
pytest==8.1.1
pytest-asyncio==0.23.6
